+++ src/plugin.hpp
@@ -25,6 +25,7 @@ using				std::unique_ptr;
 #include "RegexExp/RegexExp.hpp"
 
 #include "Tree/Tree.hpp"
+#include "TreeExp/TreeExp.hpp"
 
 #include "Gbu/Gbu.hpp"
 #include "Pkm/Pkm.hpp"
@@ -117,6 +118,7 @@ extern Model*		modelRegexCondensed;
 extern Model*		modelRegexExp;
 
 extern Model*		modelTree;
+extern Model*		modelTreeExp;
 
 extern Model*		modelGbu;
 extern Model*		modelPkm;