+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Regex
 	- [x] Fixed clock `*` modulator
 	- [x] Change cursor color for dark room (bloom light ?)
+- Tree
+	- [x] Added branch mutation chance (context menu)
+- [ ] New **Tree Exp** expander module !
 
 ## 2.0.10 (feb 26, 2024)